changes to collector service module
