+++ src/terrama2/collector/Collector.cpp
@@ -87,16 +87,16 @@ void terrama2::collector::Collector::collectAsThread(const DataSetTimerPtr datas
   //already locked by Collector::collect, lock_guard just to release when finished
   std::lock_guard<std::mutex> lock(mutex_, std::adopt_lock);
 
-  core::DataSet localDataSet = datasetTimer->dataSet();
+  core::DataSet dataSet = datasetTimer->dataSet();
 
-  if(localDataSet.dataSetItems().empty())
+  if(dataSet.dataSetItems().empty())
   {
     //TODO: LOG empty dataset
     return;
   }
 
   //aquire all data
-  for(auto& dataSetItem : localDataSet.dataSetItems())
+  for(auto& dataSetItem : dataSet.dataSetItems())
   {
 
     try
